# social-api/openapi/pipelines.yaml
...
openapi: 3.1.0
info:
  title: Pipelines HTTP API
  description: |
    REST surface for pipeline definitions, runs, metrics, and the user
    search used by the pipeline editor's UserPicker. Part of `social-api`.

    Phase 1 endpoints back onto in-memory stubs; the wire contracts are
    intentionally stable so Phase 4 can swap in distributed-core's
    PipelineModule without touching clients. See `PIPELINES_PLAN.md §17`.
  version: 1.0.0
  contact:
    name: websocket-gateway
    url: https://github.com/hoehn/websocket-gateway
servers:
  - url: /api
    description: social-api mount (Express app mounts the router at /api)
security:
  - bearerAuth: []
...
paths:
...
  /pipelines/defs:
    get:
      tags: [pipelines]
      summary: List pipeline definitions for the caller
      operationId: listPipelineDefinitions
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [pipelines]
                properties:
                  pipelines:
                    type: array
                    items:
                      $ref: '#/components/schemas/PipelineDefinition'
        '401': { $ref: '#/components/responses/Unauthorized' }
  /pipelines/defs/{pipelineId}:
    parameters:
      - $ref: '#/components/parameters/PipelineIdPath'
    get:
      tags: [pipelines]
      summary: Get one pipeline definition
      operationId: getPipelineDefinition
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineDefinition'
        '404': { $ref: '#/components/responses/NotFound' }
        '401': { $ref: '#/components/responses/Unauthorized' }

    put:
      tags: [pipelines]
      summary: Upsert a pipeline definition
      description: |
        Body `id` must match the `:pipelineId` path param; mismatch returns
        `400 invalid_body`. Creates if absent, replaces if present.
      operationId: upsertPipelineDefinition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PipelineDefinition'
      responses:
        '200':
          description: Upserted definition (echoed body)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineDefinition'
        '400':
          description: Body missing or `body.id` does not match `:pipelineId`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401': { $ref: '#/components/responses/Unauthorized' }

    delete:
      tags: [pipelines]
      summary: Delete a pipeline definition
      operationId: deletePipelineDefinition
      responses:
        '204':
          description: Deleted (or no-op if absent)
        '401': { $ref: '#/components/responses/Unauthorized' }
  /pipelines/defs/{pipelineId}/publish:
    parameters:
      - $ref: '#/components/parameters/PipelineIdPath'
    post:
      tags: [pipelines]
      summary: Publish a pipeline definition
      description: |
        Bumps `version` by 1, sets `status = 'published'`, and snapshots
        `publishedVersion` to the new `version`. Idempotent in the sense
        that each call bumps, but no additional side effects.
      operationId: publishPipelineDefinition
      responses:
        '200':
          description: Updated definition after publish
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineDefinition'
        '404': { $ref: '#/components/responses/NotFound' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # --- Runs ----------------------------------------------------------------
...
  /profiles:
    get:
      tags: [profiles]
      summary: Search user profiles by displayName / userId substring
      description: |
        Case-insensitive substring search. Visibility-gated: private
        profiles are only returned when the requester is the owner OR is
        in a mutual-follow relationship with the candidate.
      operationId: searchProfiles
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 100
          description: Search query (1..100 chars)
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
          description: Max results (clamped server-side to [1, 50])
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [profiles]
                properties:
                  profiles:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProfileSummary'
        '400':
          description: Missing or invalid `q`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401': { $ref: '#/components/responses/Unauthorized' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Cognito ID token. Passed as `Authorization: Bearer <token>`.
        Validated by `middleware/auth.ts` against the configured user pool.
  parameters:
    PipelineIdPath:
      in: path
      name: pipelineId
      required: true
      schema:
        type: string
      description: Pipeline definition identifier (opaque string, often UUID).
...
  responses:
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
  schemas:
...
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Machine-readable error code (e.g. `not_found`, `invalid_body`).
        message:
          type: string
          description: Optional human-readable detail.

    # -----------------------------------------------------------------
    # Trigger run request/response
    # -----------------------------------------------------------------
...
    ProfileSummary:
      type: object
      required: [userId, displayName]
      properties:
        userId:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
          format: uri

    # -----------------------------------------------------------------
    # Pipeline metrics
    # -----------------------------------------------------------------
...
    NodeType:
      type: string
      enum:
        - trigger
        - llm
        - transform
        - condition
        - action
        - fork
        - join
        - approval
    TriggerType:
      type: string
      enum:
        - manual
        - document.finalize
        - document.comment
        - document.submit
        - schedule
        - webhook
    LLMProvider:
      type: string
      enum: [anthropic, bedrock]
    TransformType:
      type: string
      enum: [jsonpath, template, javascript]
    ActionType:
      type: string
      enum:
        - update-document
        - post-comment
        - notify
        - webhook
        - mcp-tool
    JoinMode:
      type: string
      enum: [all, any, n_of_m]
    JoinMergeStrategy:
      type: string
      enum: [deep-merge, array-collect, last-writer-wins]
    ApprovalTimeoutAction:
      type: string
      enum: [reject, approve, escalate]
    Approver:
      type: object
      required: [type, value]
      properties:
        type:
          type: string
          enum: [user, role]
        value:
          type: string
    TriggerNodeData:
      type: object
      required: [type, triggerType]
      properties:
        type:
          type: string
          const: trigger
        triggerType:
          $ref: '#/components/schemas/TriggerType'
        documentTypeId:
          type: string
          description: Required for `document.*` triggers.
        schedule:
          type: string
          description: Cron expression; required when `triggerType === 'schedule'`.
        webhookPath:
          type: string
          description: Required when `triggerType === 'webhook'`.
    LLMNodeData:
      type: object
      required: [type, provider, model, systemPrompt, userPromptTemplate, streaming]
      properties:
        type:
          type: string
          const: llm
        provider:
          $ref: '#/components/schemas/LLMProvider'
        model:
          type: string
        systemPrompt:
          type: string
        userPromptTemplate:
          type: string
        temperature:
          type: number
          minimum: 0
          maximum: 2
        maxTokens:
          type: integer
          minimum: 1
        streaming:
          type: boolean
    TransformNodeData:
      type: object
      required: [type, transformType, expression]
      properties:
        type:
          type: string
          const: transform
        transformType:
          $ref: '#/components/schemas/TransformType'
        expression:
          type: string
        outputKey:
          type: string
    ConditionNodeData:
      type: object
      required: [type, expression]
      properties:
        type:
          type: string
          const: condition
        expression:
          type: string
        label:
          type: string
    ActionNodeData:
      type: object
      required: [type, actionType, config]
      properties:
        type:
          type: string
          const: action
        actionType:
          $ref: '#/components/schemas/ActionType'
        config:
          type: object
          additionalProperties: true
        idempotent:
          type: boolean
        onError:
          type: string
          enum: [route-error, fail-run]
    ForkNodeData:
      type: object
      required: [type, branchCount]
      properties:
        type:
          type: string
          const: fork
        branchCount:
          type: integer
          minimum: 2
          maximum: 8
        branchLabels:
          type: array
          items:
            type: string
    JoinNodeData:
      type: object
      required: [type, mode, mergeStrategy]
      properties:
        type:
          type: string
          const: join
        mode:
          $ref: '#/components/schemas/JoinMode'
        n:
          type: integer
          minimum: 1
          description: Required when `mode === 'n_of_m'`.
        mergeStrategy:
          $ref: '#/components/schemas/JoinMergeStrategy'
    ApprovalNodeData:
      type: object
      required: [type, approvers, requiredCount]
      properties:
        type:
          type: string
          const: approval
        approvers:
          type: array
          items:
            $ref: '#/components/schemas/Approver'
        requiredCount:
          type: integer
          minimum: 1
        timeoutMs:
          type: integer
          minimum: 0
        timeoutAction:
          $ref: '#/components/schemas/ApprovalTimeoutAction'
        message:
          type: string
    NodeData:
      oneOf:
        - $ref: '#/components/schemas/TriggerNodeData'
        - $ref: '#/components/schemas/LLMNodeData'
        - $ref: '#/components/schemas/TransformNodeData'
        - $ref: '#/components/schemas/ConditionNodeData'
        - $ref: '#/components/schemas/ActionNodeData'
        - $ref: '#/components/schemas/ForkNodeData'
        - $ref: '#/components/schemas/JoinNodeData'
        - $ref: '#/components/schemas/ApprovalNodeData'
      discriminator:
        propertyName: type
        mapping:
          trigger: '#/components/schemas/TriggerNodeData'
          llm: '#/components/schemas/LLMNodeData'
          transform: '#/components/schemas/TransformNodeData'
          condition: '#/components/schemas/ConditionNodeData'
          action: '#/components/schemas/ActionNodeData'
          fork: '#/components/schemas/ForkNodeData'
          join: '#/components/schemas/JoinNodeData'
          approval: '#/components/schemas/ApprovalNodeData'

    # -----------------------------------------------------------------
    # Graph
    # -----------------------------------------------------------------
    PipelineNode:
      type: object
      required: [id, type, position, data]
      properties:
        id:
          type: string
        type:
          $ref: '#/components/schemas/NodeType'
        position:
          type: object
          required: [x, y]
          properties:
            x: { type: number }
            y: { type: number }
        data:
          $ref: '#/components/schemas/NodeData'
    PipelineEdge:
      type: object
      required: [id, source, sourceHandle, target, targetHandle]
      properties:
        id:
          type: string
        source:
          type: string
        sourceHandle:
          type: string
          description: >-
            'out', 'true'/'false', 'branch-N', 'approved'/'rejected', 'error'.
        target:
          type: string
        targetHandle:
          type: string
          description: "'in' or 'in-N' for Join."
    TriggerBinding:
      type: object
      required: [event]
      properties:
        event:
          $ref: '#/components/schemas/TriggerType'
        documentTypeId: { type: string }
        schedule: { type: string }
        webhookPath: { type: string }
    PipelineStatus:
      type: string
      enum: [draft, published]
    PipelineDefinition:
      type: object
      required:
        - id
        - name
        - version
        - status
        - nodes
        - edges
        - createdAt
        - updatedAt
        - createdBy
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
          description: Normalized (lowercase, trimmed) free-form labels.
        icon:
          type: string
        version:
          type: integer
          minimum: 0
          description: Bumped on each save.
        status:
          $ref: '#/components/schemas/PipelineStatus'
        publishedVersion:
          type: integer
          minimum: 0
        publishedSnapshot:
          $ref: '#/components/schemas/PipelineDefinitionSnapshot'
          description: |
            Deep clone of the definition at last publish (minus
            `publishedSnapshot` itself, to avoid recursion). Drives the
            VersionDiffModal — populated by `publishPipeline`.
        triggerBinding:
          $ref: '#/components/schemas/TriggerBinding'
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/PipelineNode'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/PipelineEdge'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdBy:
          type: string

    # -----------------------------------------------------------------
    # Runtime
    # -----------------------------------------------------------------
...
    PipelineDefinitionSnapshot:
      type: object
      required:
        - id
        - name
        - version
        - status
        - nodes
        - edges
        - createdAt
        - updatedAt
        - createdBy
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        tags:
          type: array
          items: { type: string }
        icon: { type: string }
        version:
          type: integer
          minimum: 0
        status:
          $ref: '#/components/schemas/PipelineStatus'
        publishedVersion:
          type: integer
          minimum: 0
        triggerBinding:
          $ref: '#/components/schemas/TriggerBinding'
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/PipelineNode'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/PipelineEdge'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdBy: { type: string }

    # -----------------------------------------------------------------
    # Wire event envelope — wraps every PipelineEventMap event as it
    # crosses the network boundary. `seq` is monotonic per-run and stable
    # across replays; clients dedupe on it (NEVER on `emittedAt`).
    # -----------------------------------------------------------------
...